agent-system/requirements.txt
llama-index
langgraph
langchain
langchain-anthropic
langchain-google-genai
langchain-ollama
python-dotenv
pandas
pydantic
PyPDF2
